+++ tests/Editors/MergeIntoIndexableTest.php
@@ -34,119 +34,40 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   DataContainers/ValueBuilders
+ * @package   DataContainers/Editors
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://code.ganbarodigital.com/php-data-containers
  */
 
-namespace GanbaroDigital\DataContainers\ValueBuilders;
+namespace GanbaroDigital\DataContainers\Editors;
 
 use PHPUnit_Framework_TestCase;
 
 /**
- * @coversDefaultClass GanbaroDigital\DataContainers\ValueBuilders\MergeIntoAssignable
+ * @coversDefaultClass GanbaroDigital\DataContainers\Editors\MergeIntoIndexable
  */
-class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
+class MergeIntoIndexableTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @covers ::fromObject
      * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
+     * @covers ::mergeKeyIntoArray
      */
-    public function testCanMergeSimpleObjects()
+    public function testCanMergeSimpleArrays()
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $ours = (object)[ 'one' => 1, 'two' => 2];
-        $theirs = (object)[ 'three' => 3, 'four' => 4];
-
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4];
-
-        // ----------------------------------------------------------------
-        // perform the change
-
-        MergeIntoAssignable::fromObject($ours, $theirs);
-
-        // ----------------------------------------------------------------
-        // test the results
-
-        $this->assertEquals($expectedResult, $ours);
-    }
-
-    /**
-     * @covers ::fromObject
-     * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
-     */
-    public function testCanMergeNestedObjects()
-    {
-        // ----------------------------------------------------------------
-        // setup your test
-
-        $ours = (object)[ 'one' => 1, 'two' => 2];
-        $theirs = (object)[ 'three' => 3, 'four' => 4, 'five' => (object)[ 1,2,3]];
-
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4, 'five' => (object)[1,2,3]];
-
-        // ----------------------------------------------------------------
-        // perform the change
-
-        MergeIntoAssignable::fromObject($ours, $theirs);
-
-        // ----------------------------------------------------------------
-        // test the results
-
-        $this->assertEquals($expectedResult, $ours);
-        $this->assertSame($theirs->five, $ours->five);
-    }
-
-    /**
-     * @covers ::fromObject
-     * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
-     */
-    public function testCanMergeIntoSimpleObjects()
-    {
-        // ----------------------------------------------------------------
-        // setup your test
-
-        $ours = (object)[ 'one' => 1, 'two' => 2, 'three' => (object)['three' => 3]];
-        $theirs = (object)[ 'three' => (object)['three' => 3, 'four' => 4]];
-
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => (object) ['three' => 3, 'four' => 4]];
-
-        // ----------------------------------------------------------------
-        // perform the change
-
-        MergeIntoAssignable::fromObject($ours, $theirs);
-
-        // ----------------------------------------------------------------
-        // test the results
-
-        $this->assertEquals($expectedResult, $ours);
-    }
-
-    /**
-     * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
-     */
-    public function testCanMergeSimpleArray()
-    {
-        // ----------------------------------------------------------------
-        // setup your test
-
-        $ours = (object)[ 'one' => 1, 'two' => 2];
+        $ours = [ 'one' => 1, 'two' => 2];
         $theirs = [ 'three' => 3, 'four' => 4];
 
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4];
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4];
 
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromArray($ours, $theirs);
+        MergeIntoIndexable::fromArray($ours, $theirs);
 
         // ----------------------------------------------------------------
         // test the results
@@ -155,24 +76,23 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers ::fromObject
      * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
+     * @covers ::mergeKeyIntoArray
      */
     public function testCanMergeNestedArrays()
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $ours = (object)[ 'one' => 1, 'two' => 2];
-        $theirs = [ 'three' => 3, 'four' => 4, 'five' => (object)[ 1,2,3]];
+        $ours = [ 'one' => 1, 'two' => 2];
+        $theirs = [ 'three' => 3, 'four' => 4, 'five' => [ 1,2,3]];
 
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4, 'five' => (object)[1,2,3]];
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4, 'five' => [1,2,3]];
 
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromArray($ours, $theirs);
+        MergeIntoIndexable::fromArray($ours, $theirs);
 
         // ----------------------------------------------------------------
         // test the results
@@ -182,22 +102,99 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::fromArray
-     * @covers ::mergeKeyIntoAssignable
+     * @covers ::mergeKeyIntoArray
      */
     public function testCanMergeIntoSimpleArrays()
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $ours = (object)[ 'one' => 1, 'two' => 2, 'three' => ['three' => 3]];
+        $ours = [ 'one' => 1, 'two' => 2, 'three' => ['three' => 3]];
         $theirs = [ 'three' => ['three' => 3, 'four' => 4]];
 
-        $expectedResult = (object)[ 'one' => 1, 'two' => 2, 'three' => ['three' => 3, 'four' => 4]];
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => ['three' => 3, 'four' => 4]];
 
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromArray($ours, $theirs);
+        MergeIntoIndexable::fromArray($ours, $theirs);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $ours);
+    }
+
+    /**
+     * @covers ::fromObject
+     * @covers ::fromArray
+     * @covers ::mergeKeyIntoArray
+     */
+    public function testCanMergeSimpleObject()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $ours = [ 'one' => 1, 'two' => 2];
+        $theirs = (object)[ 'three' => 3, 'four' => 4];
+
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4];
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        MergeIntoIndexable::fromObject($ours, $theirs);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $ours);
+    }
+
+    /**
+     * @covers ::fromObject
+     * @covers ::fromArray
+     * @covers ::mergeKeyIntoArray
+     */
+    public function testCanMergeNestedObjects()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $ours = [ 'one' => 1, 'two' => 2];
+        $theirs = (object)[ 'three' => 3, 'four' => 4, 'five' => (object)[ 1,2,3]];
+
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => 3, 'four' => 4, 'five' => (object)[1,2,3]];
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        MergeIntoIndexable::fromObject($ours, $theirs);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $ours);
+    }
+
+    /**
+     * @covers ::fromObject
+     * @covers ::mergeKeyIntoArray
+     */
+    public function testCanMergeIntoSimpleObjects()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $ours = [ 'one' => 1, 'two' => 2, 'three' => (object)['three' => 3]];
+        $theirs = (object)[ 'three' => ['three' => 3, 'four' => 4]];
+
+        $expectedResult = [ 'one' => 1, 'two' => 2, 'three' => (object)['three' => 3, 'four' => 4]];
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        MergeIntoIndexable::fromObject($ours, $theirs);
 
         // ----------------------------------------------------------------
         // test the results
@@ -220,7 +217,7 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromObject($ours, $theirs);
+        MergeIntoIndexable::fromArray($ours, $theirs);
 
         // ----------------------------------------------------------------
         // test the results
@@ -232,19 +229,19 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
     {
         return [
             [
-                (object) [
+                [
                     'one' => 1,
                     'two' => false,
                     'three' => 3.1415927,
                     'four' => 'hello, world',
                 ],
-                (object) [
+                [
                     'one' => true,
                     'two' => 'two',
                     'three' => 3,
                     'four' => 4.567,
                 ],
-                (object) [
+                [
                     'one' => true,
                     'two' => 'two',
                     'three' => 3,
@@ -258,7 +255,24 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
      * @covers ::fromArray
      * @expectedException GanbaroDigital\DataContainers\Exceptions\E4xx_UnsupportedType
      */
-    public function testThrowsExceptionIfNonAssignablePassedIntoOfArray()
+    public function testThrowsExceptionIfNonAssignablePassedIntoFromArray()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $data = false;
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        MergeIntoIndexable::fromArray($data, []);
+    }
+
+    /**
+     * @covers ::fromArray
+     * @expectedException GanbaroDigital\DataContainers\Exceptions\E4xx_UnsupportedType
+     */
+    public function testThrowsExceptionIfNonTraversablePassedIntoFromArray()
     {
         // ----------------------------------------------------------------
         // setup your test
@@ -268,26 +282,14 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromArray($data, []);
-    }
-
-    /**
-     * @covers ::fromArray
-     * @expectedException GanbaroDigital\DataContainers\Exceptions\E4xx_UnsupportedType
-     */
-    public function testThrowsExceptionIfNonTraversablePassedIntoOfArray()
-    {
-        // ----------------------------------------------------------------
-        // perform the change
-
-        MergeIntoAssignable::fromArray((object)[], false);
+        MergeIntoIndexable::fromArray($data, false);
     }
 
     /**
      * @covers ::fromObject
      * @expectedException GanbaroDigital\DataContainers\Exceptions\E4xx_UnsupportedType
      */
-    public function testThrowsExceptionIfNonObjectPassedIntoOfObject()
+    public function testThrowsExceptionIfNonObjectPassedIntoFromObject()
     {
         // ----------------------------------------------------------------
         // setup your test
@@ -297,7 +299,7 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::fromObject($data, []);
+        MergeIntoIndexable::fromObject($data, false);
     }
 
     /**
@@ -311,12 +313,12 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $obj = new MergeIntoAssignable();
+        $obj = new MergeIntoIndexable();
 
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertTrue($obj instanceof MergeIntoAssignable);
+        $this->assertTrue($obj instanceof MergeIntoIndexable);
     }
 
     /**
@@ -327,12 +329,12 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // setup your test
 
-        $ours1 = (object)[];
-        $ours2 = (object)[];
+        $ours1 = [];
+        $ours2 = [];
         $theirs1 = ["one" => 1];
         $theirs2 = (object)["two" => 2];
 
-        $obj = new MergeIntoAssignable();
+        $obj = new MergeIntoIndexable();
 
         // ----------------------------------------------------------------
         // perform the change
@@ -343,34 +345,34 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertEquals($ours1, (object)$theirs1);
-        $this->assertEquals($ours2, $theirs2);
+        $this->assertEquals($ours1, $theirs1);
+        $this->assertEquals($ours2, (array)$theirs2);
     }
 
     /**
      * @covers ::from
      */
-    public function testCanStaticallyCall()
+    public function testCanStaticallyCallStatically()
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $ours1 = (object)[];
-        $ours2 = (object)[];
+        $ours1 = [];
+        $ours2 = [];
         $theirs1 = ["one" => 1];
         $theirs2 = (object)["two" => 2];
 
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::from($ours1, $theirs1);
-        MergeIntoAssignable::from($ours2, $theirs2);
+        MergeIntoIndexable::from($ours1, $theirs1);
+        MergeIntoIndexable::from($ours2, $theirs2);
 
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertEquals($ours1, (object)$theirs1);
-        $this->assertEquals($ours2, $theirs2);
+        $this->assertEquals($ours1, $theirs1);
+        $this->assertEquals($ours2, (array)$theirs2);
     }
 
     /**
@@ -388,7 +390,7 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::from($ours, $theirs);
+        MergeIntoIndexable::from($ours, $theirs);
     }
 
     /**
@@ -401,12 +403,12 @@ class MergeIntoAssignableTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // setup your test
 
-        $data = (object)[];
+        $data = [];
 
         // ----------------------------------------------------------------
         // perform the change
 
-        MergeIntoAssignable::from($data, $theirs);
+        MergeIntoIndexable::from($data, $theirs);
     }
 
     public function provideUnsupportedTypes()